Fix delay when using modem auto detect
When using modem auto detect, the autodetector's task is added
to the scheduler's task list instead of the modem driver.
But the modem driver's delay was not forwarded to the autodetector,
so it always had a constant delay of 10ms between executing the
task's run method. This is fixed by updating the autodetector's
setDelay() from it's run() method.